Detach strategy state from Service.
The strategy also serves as a cache, so it contains state. Because a
service is stateless, this causes a bug when trying to reuse a Service
instance multiple times.

To fix this, the cluster instantiates and uses the strategy during
deployment to determine placement, allowing the service to only hold a
reference to the desired strategy, thereby keeping it state-free.